requirements.txt
pyyaml>=6.0
click>=8.1.0
requests>=2.31.0
